+++ packages/fabric/editor-core/examples/5-full-page.tsx
@@ -137,9 +137,9 @@ export default class Example extends React.Component<Props, State> {
               allowUnsupportedContent={true}
               allowPanel={true}
               allowExtension={true}
-              allowPlaceholderCursor={true}
               allowRule={true}
               allowDate={true}
+              allowTableColumnResizing={true}
               {...providers}
               media={{ provider: mediaProvider, allowMediaSingle: true }}
               placeholder="Write something..."