+++ packages/fabric/editor-core/src/editor/plugins/table/index.tsx
@@ -5,10 +5,11 @@ import {
   tableHeader,
   tableRow,
 } from '@atlaskit/editor-common';
-import { tableEditing } from 'prosemirror-tables';
+import { tableEditing, columnResizing } from 'prosemirror-tables';
 import { EditorPlugin } from '../../types';
 import { plugin, stateKey } from '../../../plugins/table';
 import hoverSelectionPlugin from './hover-selection-plugin';
+import tableColumnResizingPlugin from './table-column-resizing-plugin';
 import TableFloatingToolbar from '../../../ui/TableFloatingToolbar';
 
 const tablesPlugin: EditorPlugin = {
@@ -23,9 +24,31 @@ const tablesPlugin: EditorPlugin = {
 
   pmPlugins() {
     return [
-      { rank: 900, plugin: () => plugin() },
-      { rank: 910, plugin: () => tableEditing() },
-      { rank: 920, plugin: () => hoverSelectionPlugin },
+      {
+        rank: 900,
+        plugin: ({
+          props: { allowTableColumnResizing: allowColumnResizing },
+        }) =>
+          plugin({
+            allowColumnResizing,
+          }),
+      },
+      {
+        rank: 910,
+        plugin: ({ props }) =>
+          props.allowTableColumnResizing
+            ? columnResizing({ handleWidth: 6 })
+            : undefined,
+      },
+      {
+        rank: 920,
+        plugin: ({ props }) =>
+          props.allowTableColumnResizing
+            ? tableColumnResizingPlugin
+            : undefined,
+      },
+      { rank: 930, plugin: () => tableEditing() },
+      { rank: 940, plugin: () => hoverSelectionPlugin },
     ];
   },