+++ packages/fabric/editor-core/src/editor/ui/Appearance/FullPage.tsx
@@ -59,12 +59,7 @@ const ContentArea = styled.div`
       clear: none;
     }
   }
-
-  & .ProseMirror .table-decoration {
-    left: 0;
-  }
-
-  & .ProseMirror table {
+  & .ProseMirror .table-container table {
     margin-left: 0;
     margin-right: 0;
     width: 100%;
@@ -122,6 +117,7 @@ export default class Editor extends React.Component<
       customContentComponents,
       popupsMountPoint,
       popupsBoundariesElement,
+      popupsScrollableElement,
       disabled,
     } = this.props;
 
@@ -136,6 +132,7 @@ export default class Editor extends React.Component<
             items={primaryToolbarComponents}
             popupsMountPoint={popupsMountPoint}
             popupsBoundariesElement={popupsBoundariesElement}
+            popupsScrollableElement={popupsScrollableElement}
             disabled={!!disabled}
           />
           <MainToolbarCustomComponentsSlot>
@@ -157,6 +154,7 @@ export default class Editor extends React.Component<
               items={contentComponents}
               popupsMountPoint={popupsMountPoint}
               popupsBoundariesElement={popupsBoundariesElement}
+              popupsScrollableElement={popupsScrollableElement}
               disabled={!!disabled}
             />
           </ContentArea>