+++ packages/fabric/editor-core/src/ui/TableFloatingControls/CornerControls/styles.tsx
@@ -15,6 +15,12 @@ export const CornerContainer = styled.div`
   left: -${akEditorTableToolbarSize}px;
   width: ${akEditorTableToolbarSize}px;
   height: ${akEditorTableToolbarSize}px;
+
+  display: none;
+
+  .ProseMirror .with-controls & {
+    display: block;
+  }
 `;
 
 // tslint:disable-next-line:variable-name